% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QA_QC.R
\name{build_graphical_model}
\alias{build_graphical_model}
\title{Build a graphical model given input data}
\usage{
build_graphical_model(dat, initial, vars, pred, fams, quiet = TRUE)
}
\arguments{
\item{dat}{The data.frame that will be analyzed with an Aster model. This should be the
output from read_aster_data().}

\item{initial}{The name of the initial node in the graphical model. This should correspond
to one of the column names of \code{dat}.}

\item{vars}{A vector of the names of the nodes in the graphical model, excluding the
initial node. These should correspond to column names of \code{dat}. See details.}

\item{pred}{A vector of the names of the predecessor nodes for \code{vars}. The initial node
should be included in this vector and the names should correspond to the
column names of \code{dat}. See details.}

\item{fams}{A vector of the exponential family distributions that should be used for the
edges of the graphical model. See details.}

\item{quiet}{A boolean (TRUE/FALSE) for whether summary information about the graphical
model should be printed. Default TRUE.}
}
\value{
A list with three elements:

$vars: the names of the nodes of the graphical model

$pred: the numerical index of the predecessors for each node named in the
$vars element.

$fam: the numerical index of the exponential family distribution that will
be used to model the edge from predecessor->successor
}
\description{
Build a graphical model given input data
}
\details{
This function is mostly a convenience function for specifying a graphical
model with a given data.frame. The \code{vars} argument should be a vector of the
names of the non-initial nodes in the graphical model. The \code{pred}
argument should be a vector of the names of the predecessor node for each
node given in \code{vars}, in the same order. The \code{fams} argument shoud give the
exponential family distribution that will be used to model the
predecessor->successor edges in the graphical model.

For example, consider a data.frame that has the following columns:

Initial, Germinated, Fruits, Seeds

where \code{Initial} is the number of seeds planted in an experimental unit,
\code{Germinated} is a count of the number of seeds germinated, \code{Fruits} is a
count of the number of fruits set by all plants in the unit, and \code{Seeds} is
a count of the number of seeds collected from all fruits.

A graphical model that relates these variables could be:

\if{html}{\out{<div class="sourceCode">}}\preformatted{    Bernoulli      0-Poisson    Poisson
}\if{html}{\out{</div>}}

Initial ---> Germinated ---> Fruits ---> Seeds

Where arrows denote edges, and names above arrows denote exponential
families that describe the distribution of the successor node, conditional
on the predecessor node.

This function would be called as such:

model_spec <- build_graphical_model(data, initial="Initial",
vars=c("Germinated", "Fruits", "Seeds"),
pred=c("Initial", "Germinated", "Fruits"),
fams=c(fam_bernoilli, fam_0poi, fam_poi))

This function will throw errors if the names of the nodes or the designation
of predecessor nodes are incompatible with a valid Aster graphical model.
}
