% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_QA_QC.R
\name{read_aster_data}
\alias{read_aster_data}
\title{Read a data sheet in preparation for an Aster analysis.}
\usage{
read_aster_data(filename, initial, sep = ",")
}
\arguments{
\item{filename}{Path to the data file to read (character string)}

\item{initial}{Name of the column in the data file to use as the "Initial" node. This is
case-sensitive (character string)}

\item{sep}{Separator or delimiter character used in the data file. Comma (",") for CSV
files and tab ("\t") for TSV files. Default is "," for CSV files.}
}
\value{
A data.frame containing the data in the specified file.
}
\description{
Read a data sheet in preparation for an Aster analysis.
}
\examples{
# Read a CSV file with an initial column called "Initial"
data <- read_aster_data(filename="aster_data.csv", initial="Initial", sep=",")
# Read a TSV file with an initial column called "Root"
data <- read_aster_data(filename="aster_data.tsv", initial="Root", sep="\t")
}
